change module name & update module
